+++ ci/test.sh
@@ -36,10 +36,9 @@ pushd "workers/unity"
     dotnet run -p "${PROJECT_DIR}/.shared-ci/tools/RunUnity/RunUnity.csproj" -- \
         -batchmode \
         -projectPath "${PROJECT_DIR}/workers/unity" \
-        -runTests \
-        -testPlatform editmode \
+        -runEditorTests \
         -logfile "${PROJECT_DIR}/logs/unity-editmode-test-run.log" \
-        -testResults "${EDITMODE_TEST_RESULTS_FILE}"
+        -editorTestsResultFile "${EDITMODE_TEST_RESULTS_FILE}"
 popd
 
 cleanUnity "$(pwd)/workers/unity"
@@ -62,10 +61,9 @@ pushd "test-project"
     dotnet run -p "${PROJECT_DIR}/.shared-ci/tools/RunUnity/RunUnity.csproj" -- \
         -batchmode \
         -projectPath "${PROJECT_DIR}/test-project" \
-        -runTests \
-        -testPlatform editmode \
+        -runEditorTests \
         -logfile "${PROJECT_DIR}/logs/test-project-editmode-test-run.log" \
-        -testResults "${TEST_PROJECT_EDITMODE_TEST_RESULTS_FILE}"
+        -editorTestsResultFile "${TEST_PROJECT_EDITMODE_TEST_RESULTS_FILE}"
 popd
 
 echo "--- Testing Unity: Test Project Playmode :joystick:"